requirements.txt
google-genai
langchain-community
faiss-cpu
sentence-transformers
tensorflow
scikit-learn
numpy